dbeaver/pro#9627 Refactor CloudBeaver modules
