+++ server/bundles/io.cloudbeaver.service.auth.ldap/META-INF/MANIFEST.MF
@@ -2,13 +2,13 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Cloudbeaver Web Service - LDAP
 Bundle-Vendor: DBeaver Corp
-Bundle-SymbolicName: io.cloudbeaver.service.ldap.auth;singleton:=true
+Bundle-SymbolicName: io.cloudbeaver.service.auth.ldap;singleton:=true
 Bundle-Version: 1.0.36.qualifier
 Bundle-Release-Date: 20260921
 Bundle-RequiredExecutionEnvironment: JavaSE-21
 Bundle-ActivationPolicy: lazy
 Require-Bundle: io.cloudbeaver.model;visibility:=reexport
 Bundle-Localization: OSGI-INF/l10n/bundle
-Export-Package: io.cloudbeaver.service.ldap.auth,
- io.cloudbeaver.service.ldap.auth.ssl
-Automatic-Module-Name: io.cloudbeaver.service.ldap.auth
+Export-Package: io.cloudbeaver.service.auth.ldap,
+ io.cloudbeaver.service.auth.ldap.ssl
+Automatic-Module-Name: io.cloudbeaver.service.auth.ldap
