+++ server/bundles/io.cloudbeaver.service.auth.ldap/OSGI-INF/l10n/bundle.properties
