+++ server/bundles/io.cloudbeaver.service.auth.ldap/build.properties
