+++ server/bundles/io.cloudbeaver.service.auth.ldap/pom.xml
@@ -9,7 +9,7 @@
         <version>1.0.0-SNAPSHOT</version>
          <relativePath>../</relativePath>
     </parent>
-    <artifactId>io.cloudbeaver.service.ldap.auth</artifactId>
+    <artifactId>io.cloudbeaver.service.auth.ldap</artifactId>
     <version>1.0.36-SNAPSHOT</version>
     <packaging>eclipse-plugin</packaging>
