+++ server/bundles/io.cloudbeaver.service.auth.ldap/src/io/cloudbeaver/service/auth/ldap/ssl/LdapSslSetting.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2025 DBeaver Corp and others
+ * Copyright (C) 2010-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cloudbeaver.service.ldap.auth.ssl;
+package io.cloudbeaver.service.auth.ldap.ssl;
 
 import org.jkiss.code.Nullable;
