+++ server/product/aggregate/build-full.cmd
@@ -1,6 +1,6 @@
 @echo off
 set MAVEN_OPTS=-Xmx2048m
 
-call mvn clean verify -Dheadless-platform -T 1C
+call ..\..\..\..\dbeaver-common\mvnw.cmd clean verify -Dheadless-platform -T 1C
 
 pause
